fix (sync): wrong delivery status of Ibu
**The Cause**

1. Client (the application) does:
- ANC: event date: 15 Feb 2019
- Dokumentasi Persalinan: event date: 15 Feb 2019

2. Sync (upload):
ANC and Dokumentasi Persalinan (delivery document) will be sent, and the server will add the same time for dateCreated. Because it's a bulky transaction.

3. Sync (download), sometimes like this:
Dokumentasi Persalinan data is above ANC. Because the date is the same, so no order (sorted by), and the data can be flipped.
While in the client (application), Dokumentasi Persalinan is processed first, then ANC.
That's why the status is still ANC (EDD Passed).

**The Fix**

When there are both ANC and Dokumentasi Persalinan, with actually it's a sequence, so the status should be Dokumentasi Persalinan (delivered).

https://github.com/OpenSRP/opensrp-client-sid/issues/32